Refactor: (Feedback) 로또 번호 리스트 생성 로직 이동 (AllLottoNumber, LottoNumber -> LottoTicket)
