+++ src/test/java/lotto/LottoTicketTest.java
@@ -1,10 +1,9 @@
 package lotto;
 
 import lotto.domain.LottoNumber;
-import lotto.domain.Rank;
 import lotto.domain.LottoTicket;
+import lotto.domain.Rank;
 import lotto.domain.WinLotto;
-import lotto.exception.AlreadyExistLottoNumberException;
 import lotto.exception.IllegalLottoNumberException;
 import lotto.exception.IllegalLottoNumbersSizeException;
 import org.junit.jupiter.api.DisplayName;
@@ -12,10 +11,33 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 public class LottoTicketTest {
+
+    @Nested
+    @DisplayName("로또 생성")
+    class NewLottoTestCase {
+
+        @Test
+        @DisplayName("[성공] 1부터 45까지의 중복되지 않는 번호를 생성한다.")
+        void 중복되지_않는_번호_생성() {
+            LottoTicket numbers = new LottoTicket();
+            assertThat(numbers.get()).containsAnyElementsOf(getPossibleLottoNumbers());
+        }
+
+        private Set<LottoNumber> getPossibleLottoNumbers() {
+            return IntStream.rangeClosed(1, 45)
+                    .mapToObj(LottoNumber::new)
+                    .collect(Collectors.toSet());
+        }
+
+    }
 
     @Nested
     @DisplayName("당첨 번호와 매칭")
